+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ### Feature
 
+- Add AdEx ads [#2119](https://github.com/MyEtherWallet/MyEtherWallet/pull/2119)
 - Add in QR code for mew wallet [#2173](https://github.com/MyEtherWallet/MyEtherWallet/pull/2173)
-
 - Add telegram social icon [#2167](https://github.com/MyEtherWallet/MyEtherWallet/pull/2167)
+
+### Bug
+
+- Fix mobile menu [#1873](https://github.com/MyEtherWallet/MyEtherWallet/pull/1873)
+- Fix bcvault and coolwallet [#2175](https://github.com/MyEtherWallet/MyEtherWallet/pull/2175)
 
 ### Devop